chore: rename `typescript.ts` to `generate-types.ts`
