+++ apps/studio/lib/api/self-hosted/generate-types.ts
