+++ apps/studio/pages/api/v1/projects/[ref]/types/typescript.ts
@@ -3,7 +3,7 @@ import { NextApiRequest, NextApiResponse } from 'next'
 import { constructHeaders } from 'lib/api/apiHelpers'
 import apiWrapper from 'lib/api/apiWrapper'
 import { PgMetaDatabaseError } from 'lib/api/self-hosted/types'
-import { generateTypescriptTypes } from 'lib/api/self-hosted/typescript'
+import { generateTypescriptTypes } from 'lib/api/self-hosted/generate-types'
 import { ResponseError } from 'types'
 
 export default (req: NextApiRequest, res: NextApiResponse) =>
